requirements.txt
appnope==0.1.0
decorator==4.1.2
defusedxml==0.5.0
Django==1.11.7
django-extensions==1.9.7
djangorestframework==3.7.1
djangorestframework-xml==1.3.0
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2017.3
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
typing==3.6.2
wcwidth==0.1.7